third_party/blink/web_tests/fast/scrolling/events/scrollend-event-fired-to-window.html

<!DOCTYPE HTML>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/gesture-util.js"></script>
<style>
#targetDiv {
  width: 200px;
  height: 200px;
  overflow: scroll;
}

#innerDiv {
  width: 400px;
  height: 400px;
}
</style>

<body style="margin:0">
<div id="targetDiv">
  <div id="innerDiv">
  </div>
</div>
</body>

<script>
if (window.internals)
  internals.runtimeFlags.overscrollCustomizationEnabled = true;

var target_div = document.getElementById('targetDiv');
var rect = target_div.getBoundingClientRect();
var x = rect.left + rect.width / 2;
var y = rect.top + rect.height / 2;

function scrollUp() {
  return smoothScroll(1000, x, y, GestureSourceType.TOUCH_INPUT, "up",
      SPEED_INSTANT);
}

var scrollend_arrived = false;
function onScrollEnd(event) {
  assert_false(event.cancelable);
  // scrollend events targetting document are bubbled to the window.
  assert_true(event.bubbles);
  scrollend_arrived = true;
}
window.addEventListener("scrollend", onScrollEnd);

promise_test (async (t) => {
  // Make sure that no scrollend event is sent to target_div.
  target_div.addEventListener("scrollend",
      t.unreached_func("target_div got unexpected scrollend event."));

  await waitForCompositorCommit();

  // Scroll up on target div and wait for the doc to get scrollend event.
  await scrollUp();
  await waitFor(() => { return scrollend_arrived; },
      'Window did not receive scrollend event after scroll up on target.');
  assert_equals(target_div.scrollTop, 0);
}, 'Tests that the window gets scrollend event when no element scrolls.');
</script>
